move team generation code into simple transport
