+++ baselines/QLearning/config/config.yaml
@@ -1,5 +1,5 @@
 # experiment params
-NUM_SEEDS: 5
+NUM_SEEDS: 10
 SEED: 76
 
 # wandb params
