+++ jaxmarl/environments/mpe/simple_transport.py
@@ -359,7 +359,7 @@ class SimpleTransportMPE(SimpleMPE):
 
         return self.get_obs(state), state
 
-def main():
+def unit_test():
     key = jax.random.PRNGKey(0)
 
     # Initialize environment with default settings
@@ -441,6 +441,82 @@ def main():
     assert (state.site_quota == expected_quota).all(), f"FAIL: expected quota to be {expected_quota}, got {state.site_quota}"
     print("PASS: rewards and payload update correctly when agent 0 has payload and enters the site")
 
+def generate_teams(seed):
+    # # Dictionary to track team occurrences
+    # team_occurrences = {}
+
+    # # Check for duplicates
+    # for i, team in enumerate(teams):
+    #     # Convert the team to a sorted tuple
+    #     sorted_team = tuple(sorted(tuple(agent) for agent in team))
+        
+    #     # Track occurrences of each team
+    #     if sorted_team in team_occurrences:
+    #         team_occurrences[sorted_team].append(team)
+    #         print(i)
+    #         print(team)
+    #     else:
+    #         team_occurrences[sorted_team] = [team]
+
+    # # Print duplicates
+    # print("Duplicate teams:")
+    # for team, occurrences in team_occurrences.items():
+    #     if len(occurrences) > 1:
+    #         print(f"Team {occurrences[0]} is duplicated {len(occurrences)} times.")
+
+
+    import numpy as np
+    np.random.seed(seed)
+
+    agent_values = np.array([
+        [0.0, 0.5], [0.1, 0.4], [0.2, 0.3], [0.3, 0.2], [0.4, 0.1], [0.5, 0.0]
+    ])
+
+    # Number of teams and team size
+    num_teams = 10
+    team_size = 4
+
+    # To store unique teams
+    teams = set()
+
+    # Helper function to create a sorted tuple of the team to ensure uniqueness
+    def create_team():
+        team_indices = np.random.choice(agent_values.shape[0], team_size, replace=False)
+        team = agent_values[team_indices].tolist()
+        team = tuple(sorted(tuple(agent) for agent in team))
+        return team
+
+    # Generate 10 unique teams
+    while len(teams) < num_teams:
+        team_tuple = create_team()
+        teams.add(team_tuple)
+
+    # Convert back to array for easier handling
+    teams = np.array([list([list(agent) for agent in team]) for team in teams])
+
+    # Print the teams
+    print("Train Teams:")
+    for i, team in enumerate(teams):
+        team = [list(agent) for agent in team]
+        print(f"{team}")
+
+    # for test teams, sample cap_0 from range (0-1), set cap_1 to be (1-cap_0)
+    N_test_teams = 10
+    cap_0 = np.random.uniform(0, 1, N_test_teams*team_size).reshape(N_test_teams, team_size)
+    cap_1 = 1-cap_0
+
+    # tie caps together by agent, rearrange to shape [N_teams, N_agents, N_cap]
+    test_teams = np.round(np.stack([cap_0, cap_1]).transpose(1, 2, 0), decimals=2)
+
+    print("Test Teams:")
+    for i in range(N_test_teams):
+        test_team = test_teams[i, ...].squeeze()
+        print(test_team.flatten())
+    print()
+
 
 if __name__ == "__main__":
-    main()
+    # unit_test()
+
+    SEED = 76
+    generate_teams(SEED)
